day_06_object_detection_api/requirements.txt
# Day 06 — Object Detection API (YOLOv8)
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# YOLOv8 / Computer Vision
ultralytics>=8.1.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Utilities
loguru>=0.7.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Frontend
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
